+++ wombat/windfarm/windfarm.py
@@ -2,8 +2,11 @@
 
 import logging  # type: ignore
 import os  # type: ignore
+from itertools import chain
+from math import fsum
 
 import networkx as nx  # type: ignore
+import numpy as np
 import pandas as pd  # type: ignore
 from geopy import distance  # type: ignore
 
@@ -29,6 +32,9 @@ class Windfarm:
         self._create_turbines_and_substations()
         self._create_cables()
         self.capacity = sum(self.node_system(turb).capacity for turb in self.turbine_id)
+        self._create_substation_turbine_map()
+
+        self.repair_manager._register_windfarm(self)
 
         self.env.process(self._log_operations())
 
@@ -59,8 +65,7 @@ class Windfarm:
         windfarm_layout : str
             Filename to use for reading in the windfarm layout; must be a csv file.
         """
-        layout_path = os.path.join(self.env.data_dir, "windfarm", windfarm_layout)
-
+        layout_path = str(self.env.data_dir / "windfarm" / windfarm_layout)
         layout = (
             pd.read_csv(layout_path)
             .sort_values(by=["string", "order"])
@@ -153,6 +158,31 @@ class Windfarm:
             )
         pass
 
+    def _create_substation_turbine_map(self) -> None:
+        """Creates `substation_turbine_map`, a dictionary, that maps substation(s) to
+        the dependent turbines in the windfarm, and the weighting of each turbine in the
+        windfarm.
+        """
+        # Get all turbines dependent on each substation
+        s_t_map = {
+            s_id: list(
+                chain.from_iterable(nx.dfs_successors(self.graph, source=s_id).values())
+            )
+            for s_id in self.substation_id
+        }
+
+        # Reorient the mapping to have the turbine list and the capacity-based weighting
+        # of each turbine
+        s_t_map = {
+            s_id: dict(  # type: ignore
+                turbines=np.array(s_t_map[s_id]),
+                weights=np.array([self.node_system(t).capacity for t in s_t_map[s_id]])
+                / self.capacity,
+            )
+            for s_id in s_t_map
+        }
+        self.substation_turbine_map = s_t_map
+
     def _log_operations(self):
         """Logs the operational data for a simulation."""
         system_list = list(self.graph.nodes)
@@ -198,3 +228,38 @@ class Windfarm:
             The `System` object.
         """
         return self.graph.nodes[system_id]["system"]
+
+    @property
+    def current_availability(self) -> float:
+        """Calculates the product of all system `operating_level` variables across the
+        windfarm using the following forumation
+
+        .. math::
+            \sum{
+                OperatingLevel_{substation_{i}} *
+                \sum{OperatingLevel_{turbine_{j}} * Weight_{turbine_{j}}}
+            }
+
+        where the :math:`{OperatingLevel}` is the product of the operating level
+        of each subassembly on a given system (substation or turbine), and the
+        :math:`{Weight}` is the proportion of one turbine's capacity relative to
+        the whole windfarm.
+        """  # noqa: W605
+        operating_levels = {
+            s_id: [
+                self.node_system(t).operating_level
+                for t in self.substation_turbine_map[s_id]["turbines"]  # type: ignore
+            ]
+            for s_id in self.substation_turbine_map
+        }
+        availability = fsum(
+            [
+                self.node_system(s_id).operating_level
+                * fsum(
+                    operating_levels[s_id]
+                    * self.substation_turbine_map[s_id]["weights"]  # type: ignore
+                )
+                for s_id in self.substation_turbine_map
+            ]
+        )
+        return availability