+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [Unreleased]
+ - Downtime-based and requests-based unscheduled maintenance models have been added to
+ servicing equipment
+ - `Windfarm.current_availability` is able to account for multiple substations and is now
+ an accurate calculation of the operating level across the windfarm
+
 # 0.3.1 (2021-March-2)
 
  - Updated the the `simulation/` folder to be a single-level `core/` directory.